Refactor layout for GridableLayout
I can’t believe I didn’t think about this sooner but it is so easy. We
don’t need to manually handle `contentSize.height`, we can just
leverage from `super.collectionViewContentSize.height`. This should fix
the layouts that get cut off when we use `span` and try to compensate
for when the last row appears.

This should also be a real performance booster as the height
calculation is only done once instead of twice with the old
implementation.