CarouselSpot: Add snapping scroll behavior
This change makes `CarouselSpot` automatically snap when scrolling ended, to
center the currently focused item in its viewport. This is done by querying
the underlying `UICollectionView` for an item matching the current centerpoint
of the viewport, and conditionally iterating to the right until an item was found.

As this custom behavior is implemented, uses of `UIScrollView.isPagingEnabled` is
removed, as this behavior also covers that usecase.