CarouselSpot: Fix failing test
This patch fixes that `testCarouselSetupWithPagination` was constantly failing
on the 6.0.0 branch. Since the preferredViewSize of `CarouselSpotCell` is 88 in
both dimensions, and no item size has been specified in the model, the correct
assertion should be that the item height is 88, and that the carousel gets resized
to that height as well, matching the height of its first item.

Similarly, the assertion for header height is updated to be header height + line spacing
+ carousel height.