CarouselSpot: Support placing a page indicator as an overlay
This change replaces `Layout.pageIndicator` with `.pageIndicatorPlacement`,
and introduces a `PageIndicatorPlacement` enum for it. This enables the API
user to control whether a page control gets rendered below or on top (on the
z axis) a carousel.

This change also makes sure that an added page control gets updated with the
current page as the user scrolls through a carousel.

This is a breaking change for Spot’s JSON schema, as “layout.page-indicator”
should now be expressed as a string (with either “below” or “overlay” as a
value) instead of a Bool.