Implement ComponentManager
The ComponentManager handles the mutating operations on a Component.
This new layer of abstraction is meant offload some of the
responsibilities of the Component class and to increase testability.
The previous implementations that resided in `Component+Mutation.swift`
has been moved into a new class. `Component+Mutation` still exists so
that the public API is intact, all they do is rely the operation to the
manager.