Refactor SpotsControllerManager
Calling cache in SpotsControllerManager has been removed, this will now
be handled by the app implementation. Same principals as for
ComponentManager, which is that if you want to do any additional
preprocessing of the information before you cache it to disk, you are
free to do so.

This also reduces the amount of component operation computation as
ComponentManager has that concern. SpotsControllerManager now simply
relays the instructions and relies on ComponentManager to call the
appropriate methods that are needed to complete the operation.

Some overhead methods are removed, such as `completeUpdates` and
`finishReloading`.

Small improvements are made to some of the guard statements to improve
debug-ablility.

SpotsControllerManager will now throw assertFailures if the component
cannot be resolved.

ComponentManager has one change when it comes to which queue should be
used to perform updates in `reload`, it now uses `main` instead of
`interactive`. Those methods now update height and indexes and if you
don't do that in the main queue the application could become unstable
and crash.