Improve iOS implementation when using reload with components
afterUpdate will no longer call layout as that could cause the
component to blink.

SpotsController on iOS now uses a custom implementation of
UIRefreshControl called SpotsRefreshControl. It is a tiny subclass of
UIRefreshControl that overrides `endRefreshing()`. We do this to check
that the refresh control is refreshing before calling
`super.endRefreshing()`. If we don't do this, it could interrupted the
scrolling as `endRefreshing` manipulates the scroll view it belongs to.

SpotsControllerManager.reload(with changes ... moreItems) will not call
completion using component.updateHeight, this is already covered by
component internally.

Component.updateHeightAndIndexes will now call refreshIndexes with
completion as an argument.