Fix page indicator position
We had an issue where the page indicator did not appear on screen, this
happened because it tried to configure the page control before the
collection had gotten its size. Now the implementation has been moved
into a separate method. It is now called in the layout method which
would mean that even if the collection view would dynamically change
size because a new payload has larger items, the page control would
update with its new position.

It also renames some of the optional assignment variable to  be more
consistent with the rest of the implementations. Layout is not a good
word for the collection view layout as it might be mixed up with the
`Layout` struct.