Refactor composition implementation and improve datasource and delegate
This PR refactors the composition implementation by removing some extra
layers, mainly all the composite views (GridComposite, ListComposite).
It will now use the wrapper to achieve the same kind of behavior. But
instead of having the reusable cell prepare the item, it will now rely
on the component's `computedHeight` method to get the size for the item.

Component now has a `parentComponent` property that indicates if the
component is being displayed inside of another component. It can also
be instantiated with a parent component to open up for making the setup
programmatically.

The Delegate and DataSrouce classes have been greatly improved with a
new layer of abstraction called `ViewPreparer`. All for data sources
now share the same implementation when preparing wrappable or item
configurable views.

Composable has also been removed with all of its extensions and other
implementations.