Improve overall performance and collection view layout method
Enables `useCache` for the item prepare method in Registry. Now it will
use the cache when preparing items but not when you make a view that
should be used inside a reusable view, there you always get a new
instance when calling `make`.

Renames the remaining `clean` labels to `recreateComposites`.

Vertical collection views has gained an incredible speed bump as the
layout method was previously flawed when using larger collections. The
collection view frame was always equal to the content size which would
mean that all the views would be configured at once. This could cause a
UI freeze from 1 to 2 seconds when trying to render a collection that
has more 1000+ items. Now we have a safe guard that the collection view
can never exceed the screen size. That way it can only render what
would fit on the screen.