Fix recursion when using composition and improve composite lookup
The look up for composite components was flawed, it relied on the index
of the composite which is not always the same as the item index. This
is now refactored to filter on the `itemIndex`.

There was a bug in Component that made it recursive when using
composition. It ended up calling reload() in a loop.