Improve delegate trigger when scrolling
Refactors the implementation for triggering `didReachBeginning` and
`didReachEnd` in the SpotsController. Before we calculated when the
trigger should happened based of the current offset and the content
size. Now we rely on the indicator value which is independent of the
content size meaning that we can set a constant value. We could even
make the threshold configurable for the controller if you want to
preload data earlier.