Improve layout implementation for collection views with header
The content size of a horizontal collection view could end up getting
the wrong right margin because we didn't subtract on from count when
calculating the content size width. This is fixed now in the
GridableLayout.

This removes some commented code in SpotsController.windowDidResize.
Instead of calling layoutViews in SpotsController.viewDidLayout, we now
call layoutSubviews.

The position of the header view inside of carousels are now improved so
that you get the desired positions when using layout insets. Before it
was always stuck at the top which is not what you want if you add
insets to the top.