Fix collection view layout warnings on macOS
When resizing a window that contains collection view based components,
the layouts were outputting warnings as the item could exceed the
maximum size allowed in the layout.

To fix silence and fix these warnings, we now call `invalidateLayout()`
in `resizeVerticalCollectionView`.
We also prepare the items before calling layout, this way the items get
their new size before the collection view receives a new size.

For vertical layouts, we now set size before calling prepare and
invalidateLayout in Component+macOS+Grid.

The referenced warnings are these:

The behavior of the UICollectionViewFlowLayout is not defined because:
The item height must be less than the height of the UICollectionView
minus the section insets top and bottom values.

The behavior of the UICollectionViewFlowLayout is not defined because:
The item width must be less than the width of the UICollectionView
minus the section insets left and right values.