Update and use new features on Tailor
Runs Carthage update to get the latest frameworks, mainly targeted at
the new version of Tailor.

It refactors how we map JSON to structs, it uses the type specific
methods introduced in Tailor 2.2.x to get a better fallback than what
we had before. So now if we pass a Float where we should have a Double,
Tailor will try and convert the Float into a Double before returning
the value.