Refactor configure closure on Component to use self instead of just the view
⚠️ Breaking change ⚠️

In the static configuration closure for Component, we used to pass the
view of the component so that you could use that for styling. Now we
will pass the entire component so that you can pinpoint and identify a
specific component to apply the style that you want for that component.

The configuration closure is also moved to the end of the setup method.