Refactor ItemConfigurable and introduce DynamicSizeView protocol
ItemConfigurable's configure method no longer uses inout so that cannot
manipulate the models size anymore. Going forward the new
`DynamicSizeView` protocol will be used for making views that size
depending on the content.

The configure method has also gained a `with` label to conform to the
rest of the naming conventions that we are currently using through out
the project.

`DynamicSizeView` has one required method, `computeSize(for item: Item)
-> CGSize`. It is invoked after `configure(with item: Item)` to make
sure that the view has content before calculating the size.