Refactor GridableLayout to use data source count
When using `infiniteScrolling`, the data source count is not equal to
the model if the size of the items are not 100% of the collection view
frame.
By using the data source count we can be sure that `contentSize.width`
includes all the elements, even those that are not on the model.