Add support for Component stretching.
Adds support for stretching components. If you only have one `Component` in your collection you can enable stretching for that component, which means that it will use the parent frame as it size instead of computing the remaining size of the component based of contentOffset and contentSize. This should have a positive effect on performance as it would exist earlier and perform less computation on larger collections.

It also adds support for stretching the last component in the collection. When enabled the last component will use the remaining size of the screen.

This can be enabled globally for an application by setting either Configuration.stretchSingleComponent or Configuration.stretchLastComponent to true.

It can also be enabled per controller by setting `controller.scrollView.stretchSingleComponent`, `controller.scrollView.stretchLastComponent` to true.

The default values are both `false`. These features are available on iOS and tvOS.