Improve updateIfNeeded to perform more accurate updates
If you do update if needed you don't always want a full reload. By
relaying the method to the new `updateComponentWithModel` we can now do
more accurate updates. If an update is the most appropriate action,
that will be done, if there are more items, they will be inserted etc.
If something demands a new cell if will be reloaded.
`updateComponentWithModel` is a refactoring of what we initially had
for reloading an entire collection of components. The difference now is
that it is its own private function that both operations use.