Refactor size calculations for carousel views.
When adding layout insets, they should only apply to the collection of
items. Headers and footer should not be affected by this.

Removes an unwanted pre-processor when preparing items. It will no
longer default to use the UIScreen for iOS and tvOS.

The componentWidth calculation has also been improved to properly size
the items when using itemSpacing.