Implement MoveAlgorithm and update current diffing process
To be able to implement moving items around in a collection, we need to
make a lot of changes to the source code. One being introducing two new
algorithms for calculating the new indexes when inserting and deleting
indexes. This is handled by `MoveAlgorithm` that has two methods
called: `calculateMoveForInsertedIndexes` and
`calculateMoveForDeletedIndexes`.

Other changes included here are, `ComponentModel`'s setting proper
indexes to items when initializing models with items, both using JSON
and programmatically.

Handling the result of the algoritms is handled inside the user
interface extensions that we have for `NSTableVIew`,
`NSCollectionView`, `UITableView` and `UICollectionView`.