Improve performance in reload method by opting-out of cleanup
If the controller is empty, there is no point in trying to clean up
existing views. To save a bit of CPU cycles we not check if there are
components before performing any cleanup. If the collection of
components is empty, we skip certain steps to make it fast.

`reloadIfNeeded` is improved by removing the calls to
`controller.scrollView.layoutViews()` and
`SpotsController.componentsDidReloadComponentModels?(controller)`.
These are called inside of `reload` when the operation is done.