Implement showEmptyComponent on Layout
By default, the Component will be hidden if it does not have any items.
With `showEmptyComponent` you can now configure if you want the
`Component` to still be visible even if the model has no items.
Sometimes you want the `Component` to always be visible even if its
collection is empty. Setting `showEmptyComponent` on `Layout` addresses
that need.