Optimize ComponentFlowLayout
Adds guard inside `layoutAttributesForElements` that checks that the
collection view is larger than 1 pixel. If it is one or less, the
layout will assume that the collection view is not visible on screen.
Either it could be located above or below the current view port.

`layoutAttributesForElements` will now try and intersect the frame of
the item attribute before adding it to the attributes array. This saves
some computation further down as it will no longer have to compute
attributes for items that are not visible.