Add ComponentResolvable protocol
This adds a new protocol called `ComponentResolvable`. Its an internal
protocol that gets a set of helper methods using protocol extensions.
It is used in `DataSource` and `Delegate` to reduce the amount of guard
statements that we previously had when ever either of the two had to
access the component data.