Fix rendering issue related to .integral in SpotsScrollView
In SpotsScrollView we rely on `.integral` when setting the frame. That
can sometimes manipulate the height of the frame. To avoid jumping
frames, we now check the frame height before we call `.integral` to see
if it was originally zero before the transformation. If it was zero, we
set the frame size height to be zero.

This fixes rendering issues related to `UICollectionView` when using
multiple components in a `SpotsScrollView`.