Fix crash related to queue jumping
The previous attempts to improve the user experience resulted in a bug
when trying to update the data source. Because we set the items in a
different queue, the data source could not determine that the diff was
correct after updating the data source. Now we will create a temporary
component, prepare the new items and do the diffing based on what the
new items get as values. Previously the model data was always updated
even if there wasn't any changes which is not what we want. Now the
model data is updating inside the `updateDataSource` closure.