Refactor user interface extensions to not wrap cells
When using UITableViewCell and UICollectionView cells, you don't want
those to be wrapped in other views as they can be used directly on the
interface.

The user interface implementation will now check if the view is a cell
or not. If it is the default cell for the interface then it will
register the view instead of wrapping it. Normal views will be wrapped
inside a wrapper cell.