Set numberOfPages on page control in `afterUpdate`
We had a bug that the `numberOfPages` did not represent the number of
items if you performed a mutation on the Component. In short, they
become out-of-sync. To fix this problem, we now set the `numberOfPages`
to `model.items.count` in `afterUpdate` which is invoked after all
mutations done on the `Component`.