Remove rect optimization for horizontal components.
Horizontal components needs a larger rect to work properly when
scrolling. If we check if the attribute intersects the item we can end
up with jumpy views meaning that the item will adjust itself within
milliseconds of it being displayed when scrolling fast. This does not
happens now that we have removed the optimization.