Refactor Component+Core to be less dependent on the user interface
Because we cannot access UI elements in the main thread, the code is
now refactored to use the model instead of directly accessing
properties that are only allowed in the main thread.

The code should also be easier to read now that it has been refactored
into a switch case instead of nested if statements.