Fix recursion when using nested components inside views
Refactors the register methods on
`UICollectionView+UserInterface.swift` and
`UITableView+UserInterface.swift` to not initialize a view for
comparison. Instead it will compare the `classType` with the
corresponding dequeuable views for the user interface.

Doing this caused a recursion if the view that is being registered has
a nested component inside.

This should also increase performance as the view no longer have to be
initialized before registering. It is just types... all the way down.