Move sizeForItem to ItemManager and improve safety
Collection view flow layouts are sensitive to receiving negative
values, so this commit adds a safety feature in `sizeForItem` that
makes it so that it cannot return negative values when resolving an
items size.

The implementation has also been moved from the component extension
into the `ItemManager`, this makes it easier to test.