Refactor macOS layout implementation
This refactoring re-implements the `SpotsScrollView` behavior for
macOS. Now it will use the same kind of underlaying implementation as
iOS and tvOS. As the user scrolls the views will now go through an
sizing algoritm that will resize the views to fit the current viewport.
This is done to reduce the memory consumption when using large lists.

This also fixes a bug in `NSCollectionView` where some index paths are
left out of the rendering cycle when the size of the items diff too
much.

Here is a list of other changes.
- Fixes a main thread warning when we try to access the header and
footer view in a different queue than main.
- `NoScrollView` has gotten a more fitting name and is now called
`ComponentScrollView`.
- `ItemManager` will no reduce the size of the item if it exceeds the
parents view, this fixes warnings when a user resizes the window that
the component belongs too.
- The selection is now kept when a user scrolls a list and the item
re-appears.
- A bug has been fixed in `rebuildSubviewsInLayoutOrder` on
`SpotsContentView`, it never properly removed the subview from
`subviewsInLayoutOrder`. The method now takes an argument to properly
exclude the view that should be removed.
- In `Component`, Instead of adding the user interface to the document
view, the user interface now acts as the document view for the scroll
view.
- Added documentation headers for the classes that have been touched.