Feature collection view extension to get layout as flow layout
This adds an extension on CollectionView to return the collection
view's layout as a flow layout. Seeing as we always rely on the flow
layout for Spots to function properly we always cast the layout into a
collection view layout. This should reduce the amount of casting that
we have to do.