Add maxHeight constraint when setting new computed height
The `Component`'s view height should never exceed the size of the
enclosing scroll view - content inset. If it is larger it will use the
enclosing scroll views height as it cannot be bigger than that. See
`SpotsScrollView`'s implementation for more detail on how views are
resized during scrolling.

Before this fix was implemented, the screen would flicker as it would
set a larger size and the reduce itself to the appropriate size.