Improve code structure for in Component+macOS+HeaderFooter
Instead of resolving the view as a tuple, it will now access the
`.view` property from the `make` method.

The header view is now added to the document view instead of directly
on the scroll view. This should fix the render glitch that the header
reappears if you scroll quickly.