Add height constraint to components used inside a controller
This follows the same principal as the previous change that we made for
macOS. Now the component height can never exceed the super views frame
height as it will never be bigger than that. This constraint will help
to not allocate more objects than the screen allows. Having it set a
frame size larger than the controller allows can lead to spikes as it
will temporarily allocate more views and then be adjusted by the
`SpotsScrollView` which in turn deallocates the views again. The new
code will ensure that this kind of spike allocation shouldn't occur.