Improve tvOS implementation
To make friends with the tvOS focus engine, we now have a `UITableView`
subclass called `ComponentTableVIew`. We need a subclass to set
`canBecomeFocused` to `false`.

By default, both `UITableView` and `UICollectionView` now both have
`remembersLastFocusedIndexPath` set to `true`.

`focusedSpot` has been renamed to `focusedComponent`.

`SpotsScrollView` will no longer modify the width of the subviews when
scrolling, this open up for having views that don't occupy the full
width when using them inside a `SpotsScrollView`. The `scrollView` will
now resize to the view `view.bounds` in `viewWillLayoutSubviews`.

When all the components are setup, the property `focusedComponent` will
default to the first component in `components` on `SpotsController`.

Removed some redundant code in `layoutSubviews` in `SpotsScrollView`.

`layoutTableView` has been refactored to support `inset` from `Layout`.