Introduce ItemChange to make item updating more easy to grasp
It was hard to reason about what level of item updates we had, you'd
have to read and evaluate each statement to understand what was going
on.
Now we have a `ItemChange` enum that holds that is used in
`ItemManager`. The `ItemChange` can be generated by calling
`computeChange(between:and:)`. That will give back `.hard`, `.medium`
or `.soft`.

From the `ItemChange` docs:

/// Determines how severe the update of an item is.
///
/// - hard: The data source will have to dequque a new view.
/// - medium: The view type can be reused but the content update
changed the size.
///           The data source will have to invoke an update to get the
new size on the cell.
/// - soft: Neither the view type or size did change, new data can
cleanly be