Fix bug when you try to reload an empty Component
This fixes a bug where you try to reload items in an empty Component.
The bug occurs because the Component tries to prepare its items without
a frame. To solve this issue, we simply replace the existing Component
with the temporary component that we already create to diff the
payloads.