Optimize SpotsScrollView on tvOS
This commit optimizes the SpotsScrollView extension on tvOS. The
performance hog that we previously had was caused by all of the
components trying to scroll at once. Now it will only try to scroll
components that are not visible on screen. This is determined by
checking the content offset before trying to scroll the component.