Feature focus guide on Component
This adds a focus guide to `Component`, by enabling the focus guide,
the component can remember the last focus view when navigating between
interfaces that are built using multiple components. It also opens up
for setting the preferred focused view for an individual component. The
focus guide is disabled by default.