Add built-in support for injection
If Spots is built with DEBUG, then each component will register a
listener for `INJECTION_BUNDLE_NOTIFICATION`. What this means is that
if you use Injection for Xcode, component will now reload the visible
views to get new instances of the views that you use inside the
component. It will give you React Native-esque hot reloading
out-of-the-box.