Add needsLayout to setup() & layout() methods on Component
This improves performance when reloading large collections as the
temporary component no longer call `setNeedsLayout()` and
`layoutIfNeeded()`. The temporary components that are used for
comparing and collecting differences between the new and old payload
don't need to be rendered on screen which leads to a huge performance
improvement when working with larger collection of components. Before
this could cause a bit of UI stutter as the `Component` needs to be
prepared in the main queue. The UI stutter would happen if there are no
changes to the new and old collection.

To achieve this optimization, we simply add another variable to the
`setup` and `layout` signature, telling the component what to do. If
`needsLayout` is set to `true` (which is the default) then it would do
the normal expected behavior. Internally when crafting temporary
components it will be set to `false` in `SpotsControllerManager`.