Refactor ComponentFlowLayout to increase performance
The previous algorithm was a bit flawed when it came to performance. When calculating the previous item, it had to do a reverse lookup, this didn't scale while scrolling horizontal collections with a lot of items. With this refactoring, scrolling performance should be better as it safes a reference to the previous layout attribute instead of trying to look it up each time.